// src/app/models/layer-data-prop.ts
export class tempLayerDataProp {
    EnergyLayerID: string = '';
    DataSetID: any = '';
    DataSetName: string = '';
    Description: string = '';
    UploadedBy: string = '';
    UploadedDate: string = '';
    PublishedDate: string = '';
    Source: string = '';
    Citation: string = '';
    Tags: string = '';
    Attributes: string = '';
    IsPublic: string = '';
    PreviewImage: string = '';
    UploadFileType: string = '';
    FilePathLocation: string = '';
    FilesIncluded: string = '';
    IconType: string = '';
    StrokeThicknessPercent: number = 0;
    StrokeColor: string = '';
    FillColor: string = '';
    FillOpacity?: number;
    SizePercent: number = 0;
    Opacity: number = 0;
    IsEnabled: string = '';
    SortNumber: string = '';
    DataSetGUID: string = '';
    ModifiedDate: string = '';
    Count: string = '';
    RepresentationType: string = '';
    TableName: string = '';
    FilterValue: string = '';
    LayerTypeID: string = '';
    DBFProperties: string = '';
    DetailPanelProperties: string = '';
    IsSaveSearch: string = '';
    IsLabelVisible: string = '';
    LabelField: string = '';
    DateIS: string = '';
    ExternalIconId: string = '';
    IsIconDisabled: string = '';
    ParentDataSetID: any = '';
    PropertyTypes: string = '';
    BBox: string = '';
    IsDeleted: string = '';
    IsFolder: string = '';
    ZoomMin: string = '';
    ZoomMax: string = '';
    ParcelBufferID: string = '';
    DetailPanelPropertiesMain: string = '';
    Layerindexval: number = 1;
    TreeStatus: string = '';
    IsFromHomeLookup: boolean = false;
    isEnergyLayer: boolean = false;
    LayerType: string = '';
    EnergyParentID: number = 0;
    FeatureType: string = '';

    // constructor() {
    //     this.DataSetID = '';
    //     this.DataSetName = '';
    //     this.Description = '';
    //     this.UploadedBy = '';
    //     this.UploadedDate = '';
    //     this.PublishedDate = '';
    //     this.Source = '';
    //     this.Citation = '';
    //     this.Tags = '';
    //     this.Attributes = '';
    //     this.IsPublic = '';
    //     this.PreviewImage = '';
    //     this.UploadFileType = '';
    //     this.FilePathLocation = '';
    //     this.FilesIncluded = '';
    //     this.IconType = '';
    //     this.StrokeThicknessPercent = '';
    //     this.StrokeColor = '';
    //     this.FillColor = '';
    //     this.SizePercent = '';
    //     this.Opacity = '';
    //     this.IsEnabled = '';
    //     this.SortNumber = '';
    //     this.DataSetGUID = '';
    //     this.ModifiedDate = '';
    //     this.Count = '';
    //     this.RepresentationType = '';
    //     this.TableName = '';
    //     this.FilterValue = '';
    //     this.LayerTypeID = '';
    //     this.DBFProperties = '';
    //     this.DetailPanelProperties = '';
    //     this.IsSaveSearch = '';
    //     this.IsLabelVisible = '';
    //     this.LabelField = '';
    //     this.DateIS = '';
    //     this.ExternalIconId = '';
    //     this.IsIconDisabled = '';
    //     this.ParentDataSetID = '';
    //     this.PropertyTypes = '';
    //     this.BBox = '';
    //     this.IsDeleted = '';
    //     this.IsFolder = '';
    //     this.ZoomMin = '';
    //     this.ZoomMax = '';
    //     this.ParcelBufferID = '';
    //     this.DetailPanelPropertiesMain = '';
    //     this.Layerindexval = 1;
    //     this.TreeStatus = '';
    // }
}